add fontend

+++ MERN/Goal/backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const mongoose = require('mongoose')
 
 // @@ desc get Goals 
 // @@ route GET /api/goals 
 // @@ access privates
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -19,7 +20,8 @@ const setGoals = asyncHandler(async (req, res) => {
         throw new Error('Please add text field ')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -29,11 +31,23 @@ const setGoals = asyncHandler(async (req, res) => {
 // @@ route PUT /api/goals:id 
 // @@ access private
 const updateGoal = asyncHandler(async (req, res) => {
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(401)
+        throw new Error('Invalid goal id ')
+    }
+
     const goal = await Goal.findById(req.params.id)
 
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    // make sure user update their own goal : 
+    if (goal.user.toString() !== req.user.id) {
+        req.status(401)
+        throw new Error('USer not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
@@ -45,11 +59,22 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @@ route DELETE /api/goals:id 
 // @@ access private
 const deleteGoal = asyncHandler(async (req, res) => {
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(401)
+        throw new Error('Invalid goal id ')
+    }
+
     const goal = await Goal.findById(req.params.id)
 
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    // make sure user delete their own goal : 
+    if (goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('USer not authorized')
     }
 
     await goal.deleteOne({ _id: req.params.id })
